Refactor Alpaca market data provider: modularize request handling and error management
- Introduced internal payload coercion helpers for Alpaca responses.
- Created protocols for injectable Alpaca SDK clients to enhance type safety.
- Added request contracts and validation helpers for latest quotes, equity bars, and option chains.
- Developed lazy loading adapters for Alpaca SDK to streamline client initialization.
- Simplified the main Alpaca provider by delegating responsibilities to new modules, improving maintainability and readability.
