+++ src/option_pricing/marketdata/providers/alpaca.py
@@ -1,927 +1,14 @@
-"""Alpaca market data provider shell for equities and option chain snapshots."""
+"""Public Alpaca market data provider API."""
 
 from __future__ import annotations
 
-import os
-from collections.abc import Callable, Mapping, Sequence
-from dataclasses import dataclass
-from datetime import UTC, date, datetime, timedelta
-from importlib import import_module
-from typing import Protocol, cast
-
-from option_pricing.marketdata.config import AlpacaConfig
-from option_pricing.marketdata.errors import (
-    MarketDataProviderError,
-    MissingProviderCredentialError,
-    ProviderDataUnavailableError,
-    ProviderRequestError,
-)
-
-
-class AlpacaProviderError(MarketDataProviderError):
-    """Base error for Alpaca provider failures."""
-
-
-class AlpacaMissingCredentialsError(
-    MissingProviderCredentialError,
+from option_pricing.marketdata.providers._alpaca_client import AlpacaClient
+from option_pricing.marketdata.providers._alpaca_errors import (
+    AlpacaDataUnavailableError,
+    AlpacaMissingCredentialsError,
     AlpacaProviderError,
-):
-    """Raised when an Alpaca API key or secret key is missing."""
-
-    def __init__(
-        self,
-        env_var_name: str | None = None,
-        *,
-        credential_name: str = "credential",
-    ) -> None:
-        super().__init__(
-            env_var_name,
-            provider_name="Alpaca",
-            credential_name=credential_name,
-        )
-
-
-class AlpacaRequestError(ProviderRequestError, AlpacaProviderError):
-    """Raised when an Alpaca request fails."""
-
-    def __init__(
-        self,
-        message: str,
-        *,
-        symbols: Sequence[str] | None = None,
-    ) -> None:
-        self.symbols = tuple(symbols) if symbols is not None else None
-        suffix = f" (symbols={self.symbols!r})" if self.symbols is not None else ""
-        super().__init__(f"{message}{suffix}")
-
-
-class AlpacaDataUnavailableError(ProviderDataUnavailableError, AlpacaProviderError):
-    """Raised when Alpaca returns no usable data for the requested quote."""
-
-
-@dataclass(frozen=True, slots=True)
-class _LatestQuoteRequest:
-    symbol_or_symbols: tuple[str, ...]
-    feed: str
-
-
-@dataclass(frozen=True, slots=True)
-class _EquityBarsRequest:
-    symbol_or_symbols: tuple[str, ...]
-    start: datetime
-    end: datetime
-    timeframe: str
-    limit: int | None
-    adjustment: str | None
-    sort: str | None
-    feed: str
-    asof: str | None
-
-
-@dataclass(frozen=True, slots=True)
-class _OptionChainRequest:
-    underlying: str
-    feed: str
-    expiry_gte: date
-    expiry_lte: date
-    strike_gte: float | None
-    strike_lte: float | None
-    option_type: str | None
-    root_symbol: str | None
-    updated_since: datetime | None
-
-    @property
-    def underlying_symbol(self) -> str:
-        return self.underlying
-
-    @property
-    def expiration_date_gte(self) -> date:
-        return self.expiry_gte
-
-    @property
-    def expiration_date_lte(self) -> date:
-        return self.expiry_lte
-
-    @property
-    def strike_price_gte(self) -> float | None:
-        return self.strike_gte
-
-    @property
-    def strike_price_lte(self) -> float | None:
-        return self.strike_lte
-
-    @property
-    def type(self) -> str | None:
-        return self.option_type
-
-
-class _StockDataClient(Protocol):
-    def get_stock_latest_quote(self, request_params: object) -> object:
-        """Return latest quote data for the request."""
-
-    def get_stock_bars(self, request_params: object) -> object:
-        """Return historical stock bars for the request."""
-
-
-class _OptionDataClient(Protocol):
-    def get_option_chain(self, request_params: object) -> object:
-        """Return option chain snapshot data for the request."""
-
-
-class AlpacaClient:
-    """Small injectable client for Alpaca market data."""
-
-    def __init__(
-        self,
-        api_key: str,
-        secret_key: str,
-        config: AlpacaConfig | None = None,
-        stock_data_client: _StockDataClient | None = None,
-        option_data_client: _OptionDataClient | None = None,
-    ) -> None:
-        self.config = config or AlpacaConfig()
-        self._api_key = _clean_credential(
-            api_key,
-            credential_name="API key",
-        )
-        self._secret_key = _clean_credential(
-            secret_key,
-            credential_name="secret key",
-        )
-        self._stock_data_client = stock_data_client
-        self._option_data_client = option_data_client
-
-    @classmethod
-    def from_env(
-        cls,
-        config: AlpacaConfig | None = None,
-        stock_data_client: _StockDataClient | None = None,
-        option_data_client: _OptionDataClient | None = None,
-    ) -> AlpacaClient:
-        """Create a client from the environment variables named by ``AlpacaConfig``."""
-
-        resolved_config = config or AlpacaConfig()
-        api_key = os.environ.get(resolved_config.api_key_env)
-        if api_key is None or not api_key.strip():
-            raise AlpacaMissingCredentialsError(
-                resolved_config.api_key_env,
-                credential_name="API key",
-            )
-
-        secret_key = os.environ.get(resolved_config.secret_key_env)
-        if secret_key is None or not secret_key.strip():
-            raise AlpacaMissingCredentialsError(
-                resolved_config.secret_key_env,
-                credential_name="secret key",
-            )
-
-        return cls(
-            api_key,
-            secret_key,
-            config=resolved_config,
-            stock_data_client=stock_data_client,
-            option_data_client=option_data_client,
-        )
-
-    def get_latest_equity_quotes(
-        self,
-        symbols: str | Sequence[str],
-        *,
-        asof: object | None = None,
-    ) -> dict[str, object]:
-        """Fetch latest quote data for one or more equity symbols."""
-
-        cleaned_symbols = _clean_symbols(symbols)
-        stock_client = self._resolve_stock_data_client()
-        request = self._latest_quote_request(cleaned_symbols)
-
-        getter = getattr(stock_client, "get_stock_latest_quote", None)
-        if not callable(getter):
-            raise TypeError(
-                "Alpaca stock_data_client must provide a callable "
-                "get_stock_latest_quote method"
-            )
-
-        try:
-            response = getter(request)
-        except Exception:
-            raise AlpacaRequestError(
-                "Alpaca latest equity quote request failed",
-                symbols=cleaned_symbols,
-            ) from None
-
-        quotes = _coerce_quote_mapping(response, cleaned_symbols)
-        return {
-            "symbols": cleaned_symbols,
-            "quotes": quotes,
-            "source": "alpaca",
-            "feed": self.config.feed,
-            "asof": asof,
-        }
-
-    def get_equity_bars(
-        self,
-        symbols: str | Sequence[str],
-        *,
-        start: datetime | str,
-        end: datetime | str,
-        timeframe: str,
-        limit: int | None = None,
-        adjustment: str | None = None,
-        sort: str | None = "asc",
-        feed: str | None = None,
-        asof: str | None = None,
-    ) -> dict[str, object]:
-        """Fetch historical equity bars for one or more symbols."""
-
-        cleaned_symbols = _clean_symbols(symbols)
-        normalized_start = _normalize_datetime(start, "start")
-        normalized_end = _normalize_datetime(end, "end")
-        if normalized_start >= normalized_end:
-            raise ValueError("start must be before end for Alpaca equity bars")
-
-        cleaned_timeframe = _clean_required_text(timeframe, "timeframe")
-        cleaned_limit = _clean_limit(limit)
-        cleaned_adjustment = _clean_optional_text(adjustment, "adjustment")
-        cleaned_sort = _clean_optional_text(sort, "sort")
-        effective_feed = _clean_optional_text(feed, "feed") or self.config.feed
-        cleaned_asof = _clean_optional_text(asof, "asof")
-
-        stock_client = self._resolve_stock_data_client()
-        request = self._equity_bars_request(
-            cleaned_symbols,
-            start=normalized_start,
-            end=normalized_end,
-            timeframe=cleaned_timeframe,
-            limit=cleaned_limit,
-            adjustment=cleaned_adjustment,
-            sort=cleaned_sort,
-            feed=effective_feed,
-            asof=cleaned_asof,
-        )
-
-        getter = getattr(stock_client, "get_stock_bars", None)
-        if not callable(getter):
-            raise TypeError(
-                "Alpaca stock_data_client must provide a callable "
-                "get_stock_bars method"
-            )
-
-        try:
-            response = getter(request)
-        except Exception:
-            raise AlpacaRequestError(
-                "Alpaca equity bars request failed",
-                symbols=cleaned_symbols,
-            ) from None
-
-        return {
-            "symbols": cleaned_symbols,
-            "bars": getattr(response, "data", response),
-            "source": "alpaca",
-            "feed": effective_feed,
-            "timeframe": cleaned_timeframe,
-            "start": normalized_start,
-            "end": normalized_end,
-            "asof": cleaned_asof,
-        }
-
-    def get_option_chain(
-        self,
-        underlying: str,
-        *,
-        expiry_gte: date | str | None = None,
-        expiry_lte: date | str | None = None,
-        strike_gte: float | None = None,
-        strike_lte: float | None = None,
-        option_type: str | None = None,
-        root_symbol: str | None = None,
-        updated_since: datetime | str | None = None,
-        feed: str | None = None,
-        asof: object | None = None,
-    ) -> dict[str, object]:
-        """Fetch option chain snapshots for one underlying symbol."""
-
-        cleaned_underlying = _clean_underlying_symbol(underlying)
-        effective_feed = _clean_optional_text(feed, "feed") or self.config.feed
-        normalized_expiry_gte, normalized_expiry_lte = _option_chain_expiry_bounds(
-            expiry_gte,
-            expiry_lte,
-        )
-        normalized_strike_gte = _clean_optional_float(strike_gte, "strike_gte")
-        normalized_strike_lte = _clean_optional_float(strike_lte, "strike_lte")
-        if (
-            normalized_strike_gte is not None
-            and normalized_strike_lte is not None
-            and normalized_strike_gte > normalized_strike_lte
-        ):
-            raise ValueError("strike_gte must be less than or equal to strike_lte")
-
-        cleaned_option_type = _clean_option_type(option_type)
-        cleaned_root_symbol = _clean_optional_symbol(root_symbol, "root_symbol")
-        normalized_updated_since = _normalize_optional_datetime(
-            updated_since,
-            "updated_since",
-        )
-
-        option_client = self._resolve_option_data_client()
-        request = self._option_chain_request(
-            cleaned_underlying,
-            feed=effective_feed,
-            expiry_gte=normalized_expiry_gte,
-            expiry_lte=normalized_expiry_lte,
-            strike_gte=normalized_strike_gte,
-            strike_lte=normalized_strike_lte,
-            option_type=cleaned_option_type,
-            root_symbol=cleaned_root_symbol,
-            updated_since=normalized_updated_since,
-        )
-
-        getter = getattr(option_client, "get_option_chain", None)
-        if not callable(getter):
-            raise TypeError(
-                "Alpaca option_data_client must provide a callable "
-                "get_option_chain method"
-            )
-
-        try:
-            response = getter(request)
-        except Exception:
-            raise AlpacaRequestError(
-                "Alpaca option chain request failed",
-                symbols=(cleaned_underlying,),
-            ) from None
-
-        request_metadata = _option_chain_request_metadata(
-            cleaned_underlying,
-            feed=effective_feed,
-            expiry_gte=normalized_expiry_gte,
-            expiry_lte=normalized_expiry_lte,
-            strike_gte=normalized_strike_gte,
-            strike_lte=normalized_strike_lte,
-            option_type=cleaned_option_type,
-            root_symbol=cleaned_root_symbol,
-            updated_since=normalized_updated_since,
-        )
-        return {
-            "underlying": cleaned_underlying,
-            "contracts": getattr(response, "data", response),
-            "source": "alpaca",
-            "feed": effective_feed,
-            "request": request_metadata,
-            "asof": asof,
-        }
-
-    def _resolve_stock_data_client(self) -> object:
-        if self._stock_data_client is not None:
-            return self._stock_data_client
-
-        try:
-            client_cls, _, _ = _alpaca_sdk_objects()
-            return client_cls(
-                api_key=self._api_key,
-                secret_key=self._secret_key,
-                sandbox=self.config.sandbox,
-            )
-        except AlpacaProviderError:
-            raise
-        except Exception:
-            raise AlpacaRequestError(
-                "Could not construct Alpaca stock data client"
-            ) from None
-
-    def _resolve_option_data_client(self) -> object:
-        if self._option_data_client is not None:
-            return self._option_data_client
-
-        try:
-            client_cls, _, _, _ = _alpaca_option_sdk_objects()
-            return client_cls(
-                api_key=self._api_key,
-                secret_key=self._secret_key,
-                sandbox=self.config.sandbox,
-            )
-        except AlpacaProviderError:
-            raise
-        except Exception:
-            raise AlpacaRequestError(
-                "Could not construct Alpaca option data client"
-            ) from None
-
-    def _latest_quote_request(self, symbols: tuple[str, ...]) -> object:
-        if self._stock_data_client is not None:
-            return _LatestQuoteRequest(
-                symbol_or_symbols=symbols,
-                feed=self.config.feed,
-            )
-
-        try:
-            _, request_cls, data_feed_factory = _alpaca_sdk_objects()
-            feed = data_feed_factory(self.config.feed)
-            return request_cls(symbol_or_symbols=list(symbols), feed=feed)
-        except AlpacaProviderError:
-            raise
-        except Exception:
-            raise AlpacaRequestError(
-                "Could not construct Alpaca latest quote request",
-                symbols=symbols,
-            ) from None
-
-    def _equity_bars_request(
-        self,
-        symbols: tuple[str, ...],
-        *,
-        start: datetime,
-        end: datetime,
-        timeframe: str,
-        limit: int | None,
-        adjustment: str | None,
-        sort: str | None,
-        feed: str,
-        asof: str | None,
-    ) -> object:
-        if self._stock_data_client is not None:
-            return _EquityBarsRequest(
-                symbol_or_symbols=symbols,
-                start=start,
-                end=end,
-                timeframe=timeframe,
-                limit=limit,
-                adjustment=adjustment,
-                sort=sort,
-                feed=feed,
-                asof=asof,
-            )
-
-        try:
-            (
-                request_cls,
-                timeframe_factory,
-                data_feed_factory,
-                adjustment_factory,
-                sort_factory,
-            ) = _alpaca_bars_sdk_objects()
-            request_kwargs: dict[str, object] = {
-                "symbol_or_symbols": list(symbols),
-                "start": start,
-                "end": end,
-                "timeframe": _alpaca_timeframe(timeframe, timeframe_factory),
-                "feed": data_feed_factory(feed),
-            }
-            if limit is not None:
-                request_kwargs["limit"] = limit
-            if adjustment is not None:
-                request_kwargs["adjustment"] = _optional_sdk_enum(
-                    adjustment,
-                    adjustment_factory,
-                )
-            if sort is not None:
-                request_kwargs["sort"] = _optional_sdk_enum(sort, sort_factory)
-            if asof is not None:
-                request_kwargs["asof"] = asof
-            return request_cls(**request_kwargs)
-        except AlpacaProviderError:
-            raise
-        except Exception:
-            raise AlpacaRequestError(
-                "Could not construct Alpaca equity bars request",
-                symbols=symbols,
-            ) from None
-
-    def _option_chain_request(
-        self,
-        underlying: str,
-        *,
-        feed: str,
-        expiry_gte: date,
-        expiry_lte: date,
-        strike_gte: float | None,
-        strike_lte: float | None,
-        option_type: str | None,
-        root_symbol: str | None,
-        updated_since: datetime | None,
-    ) -> object:
-        if self._option_data_client is not None:
-            return _OptionChainRequest(
-                underlying=underlying,
-                feed=feed,
-                expiry_gte=expiry_gte,
-                expiry_lte=expiry_lte,
-                strike_gte=strike_gte,
-                strike_lte=strike_lte,
-                option_type=option_type,
-                root_symbol=root_symbol,
-                updated_since=updated_since,
-            )
-
-        try:
-            (
-                _,
-                request_cls,
-                options_feed_factory,
-                contract_type_factory,
-            ) = _alpaca_option_sdk_objects()
-            request_kwargs: dict[str, object] = {
-                "underlying_symbol": underlying,
-                "feed": _optional_sdk_enum(feed, options_feed_factory),
-                "expiration_date_gte": expiry_gte,
-                "expiration_date_lte": expiry_lte,
-            }
-            if strike_gte is not None:
-                request_kwargs["strike_price_gte"] = strike_gte
-            if strike_lte is not None:
-                request_kwargs["strike_price_lte"] = strike_lte
-            if option_type is not None:
-                request_kwargs["type"] = _optional_sdk_enum(
-                    option_type,
-                    contract_type_factory,
-                )
-            if root_symbol is not None:
-                request_kwargs["root_symbol"] = root_symbol
-            if updated_since is not None:
-                request_kwargs["updated_since"] = updated_since
-            return request_cls(**request_kwargs)
-        except AlpacaProviderError:
-            raise
-        except Exception:
-            raise AlpacaRequestError(
-                "Could not construct Alpaca option chain request",
-                symbols=(underlying,),
-            ) from None
-
-    def __repr__(self) -> str:
-        stock_injected = self._stock_data_client is not None
-        option_injected = self._option_data_client is not None
-        return (
-            "AlpacaClient("
-            f"config={self.config!r}, "
-            "api_key=<redacted>, "
-            "secret_key=<redacted>, "
-            f"stock_data_client_injected={stock_injected!r}, "
-            f"option_data_client_injected={option_injected!r}"
-            ")"
-        )
-
-
-def _alpaca_sdk_objects() -> (
-    tuple[Callable[..., object], Callable[..., object], Callable[[str], object]]
-):
-    try:
-        historical_module = import_module("alpaca.data.historical")
-        requests_module = import_module("alpaca.data.requests")
-        enums_module = import_module("alpaca.data.enums")
-    except ModuleNotFoundError:
-        raise AlpacaRequestError(
-            "alpaca-py is required to fetch Alpaca latest equity quotes"
-        ) from None
-
-    client_cls = getattr(historical_module, "StockHistoricalDataClient", None)
-    request_cls = getattr(requests_module, "StockLatestQuoteRequest", None)
-    data_feed_factory = getattr(enums_module, "DataFeed", None)
-
-    if not callable(client_cls):
-        raise AlpacaRequestError("alpaca-py StockHistoricalDataClient is unavailable")
-    if not callable(request_cls):
-        raise AlpacaRequestError("alpaca-py StockLatestQuoteRequest is unavailable")
-    if not callable(data_feed_factory):
-        raise AlpacaRequestError("alpaca-py DataFeed enum is unavailable")
-
-    return (
-        cast(Callable[..., object], client_cls),
-        cast(Callable[..., object], request_cls),
-        cast(Callable[[str], object], data_feed_factory),
-    )
-
-
-def _alpaca_bars_sdk_objects() -> tuple[
-    Callable[..., object],
-    Callable[..., object],
-    Callable[[str], object],
-    Callable[[str], object] | None,
-    Callable[[str], object] | None,
-]:
-    try:
-        requests_module = import_module("alpaca.data.requests")
-        enums_module = import_module("alpaca.data.enums")
-        timeframe_module = import_module("alpaca.data.timeframe")
-    except ModuleNotFoundError:
-        raise AlpacaRequestError(
-            "alpaca-py is required to fetch Alpaca equity bars"
-        ) from None
-
-    request_cls = getattr(requests_module, "StockBarsRequest", None)
-    timeframe_factory = getattr(timeframe_module, "TimeFrame", None)
-    data_feed_factory = getattr(enums_module, "DataFeed", None)
-    adjustment_factory = getattr(enums_module, "Adjustment", None)
-    sort_factory = getattr(enums_module, "Sort", None)
-
-    if not callable(request_cls):
-        raise AlpacaRequestError("alpaca-py StockBarsRequest is unavailable")
-    if timeframe_factory is None:
-        raise AlpacaRequestError("alpaca-py TimeFrame is unavailable")
-    if not callable(data_feed_factory):
-        raise AlpacaRequestError("alpaca-py DataFeed enum is unavailable")
-
-    return (
-        cast(Callable[..., object], request_cls),
-        cast(Callable[..., object], timeframe_factory),
-        cast(Callable[[str], object], data_feed_factory),
-        cast(Callable[[str], object] | None, adjustment_factory),
-        cast(Callable[[str], object] | None, sort_factory),
-    )
-
-
-def _alpaca_option_sdk_objects() -> tuple[
-    Callable[..., object],
-    Callable[..., object],
-    Callable[[str], object] | None,
-    Callable[[str], object] | None,
-]:
-    try:
-        historical_module = import_module("alpaca.data.historical.option")
-        requests_module = import_module("alpaca.data.requests")
-        enums_module = import_module("alpaca.data.enums")
-    except ModuleNotFoundError:
-        raise AlpacaRequestError(
-            "alpaca-py is required to fetch Alpaca option chains"
-        ) from None
-
-    client_cls = getattr(historical_module, "OptionHistoricalDataClient", None)
-    request_cls = getattr(requests_module, "OptionChainRequest", None)
-    options_feed_factory = getattr(enums_module, "OptionsFeed", None) or getattr(
-        requests_module,
-        "OptionsFeed",
-        None,
-    )
-    contract_type_factory = getattr(enums_module, "ContractType", None) or getattr(
-        requests_module,
-        "ContractType",
-        None,
-    )
-    if contract_type_factory is None:
-        try:
-            trading_enums_module = import_module("alpaca.trading.enums")
-        except ModuleNotFoundError:
-            trading_enums_module = None
-        if trading_enums_module is not None:
-            contract_type_factory = getattr(
-                trading_enums_module,
-                "ContractType",
-                None,
-            )
-    if not callable(options_feed_factory):
-        options_feed_factory = None
-    if not callable(contract_type_factory):
-        contract_type_factory = None
-
-    if not callable(client_cls):
-        raise AlpacaRequestError("alpaca-py OptionHistoricalDataClient is unavailable")
-    if not callable(request_cls):
-        raise AlpacaRequestError("alpaca-py OptionChainRequest is unavailable")
-
-    return (
-        cast(Callable[..., object], client_cls),
-        cast(Callable[..., object], request_cls),
-        cast(Callable[[str], object] | None, options_feed_factory),
-        cast(Callable[[str], object] | None, contract_type_factory),
-    )
-
-
-def _alpaca_timeframe(value: str, timeframe_factory: Callable[..., object]) -> object:
-    normalized = value.strip().lower().replace("_", "").replace(" ", "")
-    timeframe_attrs = {
-        "1min": "Minute",
-        "1minute": "Minute",
-        "minute": "Minute",
-        "min": "Minute",
-        "1hour": "Hour",
-        "hour": "Hour",
-        "1h": "Hour",
-        "1day": "Day",
-        "day": "Day",
-        "daily": "Day",
-        "1d": "Day",
-        "1week": "Week",
-        "week": "Week",
-        "weekly": "Week",
-        "1month": "Month",
-        "month": "Month",
-        "monthly": "Month",
-    }
-    attr_name = timeframe_attrs.get(normalized)
-    if attr_name is not None and hasattr(timeframe_factory, attr_name):
-        return getattr(timeframe_factory, attr_name)
-
-    try:
-        return timeframe_factory(value)
-    except Exception:
-        return value
-
-
-def _optional_sdk_enum(
-    value: str,
-    factory: Callable[[str], object] | None,
-) -> object:
-    if factory is None:
-        return value
-    return factory(value)
-
-
-def _coerce_quote_mapping(
-    response: object,
-    symbols: tuple[str, ...],
-) -> dict[str, object]:
-    data = getattr(response, "data", response)
-    if not isinstance(data, Mapping):
-        raise AlpacaDataUnavailableError(
-            "Alpaca latest equity quote response must be a mapping"
-        )
-
-    quotes = {str(symbol): quote for symbol, quote in data.items()}
-    missing = [symbol for symbol in symbols if symbol not in quotes]
-    if missing:
-        raise AlpacaDataUnavailableError(
-            f"Alpaca latest equity quote response is missing symbols: {missing}"
-        )
-    return quotes
-
-
-def _clean_symbols(symbols: str | Sequence[str]) -> tuple[str, ...]:
-    raw_symbols: tuple[str, ...]
-    if isinstance(symbols, str):
-        raw_symbols = (symbols,)
-    else:
-        raw_symbols = tuple(symbols)
-
-    cleaned = tuple(_clean_symbol(symbol) for symbol in raw_symbols)
-    if not cleaned:
-        raise ValueError("symbols must contain at least one symbol")
-    if len(set(cleaned)) != len(cleaned):
-        raise ValueError("symbols must not contain duplicates")
-    return cleaned
-
-
-def _option_chain_expiry_bounds(
-    expiry_gte: date | str | None,
-    expiry_lte: date | str | None,
-) -> tuple[date, date]:
-    current_date = datetime.now(UTC).date()
-    lower = (
-        current_date
-        if expiry_gte is None
-        else _normalize_date(expiry_gte, "expiry_gte")
-    )
-    upper = (
-        current_date + timedelta(days=45)
-        if expiry_lte is None
-        else _normalize_date(expiry_lte, "expiry_lte")
-    )
-    if lower > upper:
-        raise ValueError("expiry_gte must be less than or equal to expiry_lte")
-    return lower, upper
-
-
-def _normalize_date(value: date | str, field_name: str) -> date:
-    if isinstance(value, datetime):
-        if value.tzinfo is None:
-            return value.date()
-        return value.astimezone(UTC).date()
-    if isinstance(value, date):
-        return value
-    if isinstance(value, str):
-        cleaned = _clean_required_text(value, field_name)
-        try:
-            return date.fromisoformat(cleaned)
-        except ValueError as exc:
-            raise ValueError(f"{field_name} must be an ISO date string") from exc
-    raise TypeError(f"{field_name} must be a date or ISO date string")
-
-
-def _normalize_datetime(value: datetime | str, field_name: str) -> datetime:
-    if isinstance(value, datetime):
-        parsed = value
-    elif isinstance(value, str):
-        cleaned = _clean_required_text(value, field_name)
-        try:
-            parsed = datetime.fromisoformat(cleaned.replace("Z", "+00:00"))
-        except ValueError as exc:
-            raise ValueError(f"{field_name} must be an ISO datetime string") from exc
-    else:
-        raise TypeError(f"{field_name} must be a datetime or ISO datetime string")
-
-    if parsed.tzinfo is None:
-        return parsed.replace(tzinfo=UTC)
-    return parsed.astimezone(UTC)
-
-
-def _normalize_optional_datetime(
-    value: datetime | str | None,
-    field_name: str,
-) -> datetime | None:
-    if value is None:
-        return None
-    return _normalize_datetime(value, field_name)
-
-
-def _clean_required_text(value: str, field_name: str) -> str:
-    if not isinstance(value, str):
-        raise TypeError(f"{field_name} must be a string")
-    cleaned = value.strip()
-    if not cleaned:
-        raise ValueError(f"{field_name} must be a non-empty string")
-    return cleaned
-
-
-def _clean_optional_text(value: str | None, field_name: str) -> str | None:
-    if value is None:
-        return None
-    return _clean_required_text(value, field_name)
-
-
-def _clean_optional_symbol(value: str | None, field_name: str) -> str | None:
-    if value is None:
-        return None
-    return _clean_required_text(value, field_name).upper()
-
-
-def _clean_underlying_symbol(value: str) -> str:
-    return _clean_required_text(value, "underlying").upper()
-
-
-def _clean_optional_float(value: float | None, field_name: str) -> float | None:
-    if value is None:
-        return None
-    if isinstance(value, bool):
-        raise TypeError(f"{field_name} must be numeric")
-    try:
-        cleaned = float(value)
-    except (TypeError, ValueError) as exc:
-        raise TypeError(f"{field_name} must be numeric") from exc
-    if not cleaned == cleaned or cleaned in (float("inf"), float("-inf")):
-        raise ValueError(f"{field_name} must be finite")
-    return cleaned
-
-
-def _clean_option_type(value: str | None) -> str | None:
-    if value is None:
-        return None
-    cleaned = _clean_required_text(value, "option_type").lower()
-    if cleaned not in {"call", "put"}:
-        raise ValueError("option_type must be either 'call' or 'put'")
-    return cleaned
-
-
-def _option_chain_request_metadata(
-    underlying: str,
-    *,
-    feed: str,
-    expiry_gte: date,
-    expiry_lte: date,
-    strike_gte: float | None,
-    strike_lte: float | None,
-    option_type: str | None,
-    root_symbol: str | None,
-    updated_since: datetime | None,
-) -> dict[str, object]:
-    return {
-        "underlying": underlying,
-        "feed": feed,
-        "expiry_gte": expiry_gte,
-        "expiry_lte": expiry_lte,
-        "strike_gte": strike_gte,
-        "strike_lte": strike_lte,
-        "option_type": option_type,
-        "root_symbol": root_symbol,
-        "updated_since": updated_since,
-    }
-
-
-def _clean_limit(value: int | None) -> int | None:
-    if value is None:
-        return None
-    if not isinstance(value, int) or isinstance(value, bool):
-        raise TypeError("limit must be an integer")
-    if value <= 0:
-        raise ValueError("limit must be a positive integer")
-    return value
-
-
-def _clean_symbol(value: str) -> str:
-    if not isinstance(value, str):
-        raise TypeError("symbols must contain strings")
-    cleaned = value.strip().upper()
-    if not cleaned:
-        raise ValueError("symbols must contain non-empty strings")
-    return cleaned
-
-
-def _clean_credential(value: str, *, credential_name: str) -> str:
-    if not isinstance(value, str):
-        raise TypeError(f"Alpaca {credential_name} must be a string")
-    if not value.strip():
-        raise AlpacaMissingCredentialsError(credential_name=credential_name)
-    return value
-
+    AlpacaRequestError,
+)
 
 __all__ = [
     "AlpacaClient",
